more minor testing tweaks

+++ tests/take_test.py
@@ -252,7 +252,8 @@ if pytest_return == 0:
     else: 
         log.info('All tests successful: tear down test.')
         log.debug('')
-        shutil.rmtree(candidate_spec['repos_dir'])
+        if candidate_spec['repos_dir'].exists():
+            shutil.rmtree(candidate_spec['repos_dir'])
         shutil.rmtree(candidate_spec['test_dir'])
     log.debug('')
 else:
